requirements.txt
numpy
Faker
flask
flask_restful
flask_cors
pandas
plotly
seaborn
streamlit
scikit-learn==1.0.2
joblib
Unidecode
jaro-winkler